requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Cerberus==1.3.4
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
numpy==1.21.2
pandas==1.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.3.0
